Move static methods into jdt.core.manipulation.
- Move static methods from ChainCompletionTemplateBuilder (jdt.ui)
  into ChainElement (jdt.core.manipulation)
